Introduce SecAuditLogSkipNoauditlog directive
The 'noauditlog' actionset item is nebulous at best, and its
useage can be counteruitive at times. Existence of this item
will help the logging engine determine if the transaction is
worth sending to the audit log; it does not apply on a per-rule
basis. Currently all rules that match before the transaction is
logged are unconditionally added to the msr->matched_rules array,
and subsequently are written to the audit log.

This commit introduces a new configuration directive that would
prevent rules that explicitly set 'noauditlog' from being written
to the audit log, by skipping the apr_array_push call when the
rule matches. This can help reduce audit log noise by skipping
over rules that do not need to be logged (such as SecAction rules
that initialize persistent storage, etc).

By default this behavior is not enabled, so no change to the
existing audit logging logic is put in place without explicit
configuration.

Syntax:
SecAuditLogSkipNoauditlog (On|Off)

Default:
SecAuditLogSkipNoauditlog Off

Example Usage:
SecAuditLogSkipNoauditlog On